StartMenu 1.3 (changed the uint8(c++) and byte(bp) into int32(c++) and integer(bp), worked for a bit but now it stopped working again idk why)

+++ Source/ALevel/CellGeneratorAndController.cpp
@@ -20,7 +20,7 @@ void ACellGeneratorAndController::BPtestFunction()
 }
 
 
-void ACellGeneratorAndController::GenerateCells(uint8 PassedInHeight, uint8 PassedInWidth)
+void ACellGeneratorAndController::GenerateCells(int32 PassedInHeight, int32 PassedInWidth)
 {
 	/*
 	UWorld* World = GetWorld();
